refactor(FileLocationDescription): typo & add link
